requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-ollama==0.2.3
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
ollama==0.4.7
openai==1.63.0
openai-whisper==20240930
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.15.2
setuptools==75.8.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.38
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0
